--- a/kani-compiler/src/kani_middle/attributes.rs
+++ b/kani-compiler/src/kani_middle/attributes.rs
@@ -78,30 +78,11 @@ pub fn is_test_harness_description(tcx: TyCtxt, def_id: DefId) -> bool {
     tcx.sess.contains_name(attrs, rustc_span::symbol::sym::rustc_test_marker)
 }
 
-/// Is this the closure inside of a test description const (i.e. macro expanded from a `#[test]`)?
-///
-/// We're trying to detect the closure (`||`) inside code like:
-///
-/// ```ignore
-/// #[rustc_test_marker]
-/// pub const check_2: test::TestDescAndFn = test::TestDescAndFn {
-///     desc: ...,
-///     testfn: test::StaticTestFn(|| test::assert_test_result(check_2())),
-/// };
-/// ```
-pub fn is_test_harness_closure(tcx: TyCtxt, def_id: DefId) -> bool {
-    if !def_id.is_local() {
-        return false;
-    }
-
-    let local_def_id = def_id.expect_local();
-    let hir_id = tcx.hir().local_def_id_to_hir_id(local_def_id);
-
-    // The parent item of the closure appears to reliably be the `const` declaration item.
-    let parent_id = tcx.hir().get_parent_item(hir_id);
-    let parent_def_id = parent_id.to_def_id();
-
-    is_test_harness_description(tcx, parent_def_id)
+/// Extract the test harness name from the `#[rustc_test_maker]`
+pub fn test_harness_name(tcx: TyCtxt, def_id: DefId) -> String {
+    let attrs = tcx.get_attrs_unchecked(def_id);
+    let marker = tcx.sess.find_by_name(attrs, rustc_span::symbol::sym::rustc_test_marker).unwrap();
+    parse_str_value(&marker).unwrap()
 }
 
 /// Extract all Kani attributes for a given `def_id` if any exists.
@@ -490,6 +471,16 @@ fn parse_key_values(attr: &Attribute) -> Result<BTreeMap<String, String>, String
             )),
         })
         .collect()
+}
+
+/// Extracts the string value argument from the attribute provided.
+///
+/// For attributes with the following format, this will return a string that represents "VALUE".
+/// - `#[attribute = "VALUE"]`
+fn parse_str_value(attr: &Attribute) -> Option<String> {
+    // Vector of meta items , that contain the arguments given the attribute
+    let value = attr.value_str();
+    value.map(|sym| sym.to_string())
 }
 
 /// If the attribute is named `kanitool::name`, this extracts `name`