--- a/kani-compiler/src/kani_middle/metadata.rs
+++ b/kani-compiler/src/kani_middle/metadata.rs
@@ -16,7 +16,14 @@ use super::{attributes::extract_harness_attributes, SourceLocation};
 pub fn gen_proof_metadata(tcx: TyCtxt, def_id: DefId, base_name: &Path) -> HarnessMetadata {
     let attributes = extract_harness_attributes(tcx, def_id);
     let pretty_name = tcx.def_path_str(def_id);
-    let mangled_name = tcx.symbol_name(Instance::mono(tcx, def_id)).to_string();
+    // Main function a special case in order to support `--function main`
+    // TODO: Get rid of this: https://github.com/model-checking/kani/issues/2129
+    let mangled_name = if pretty_name == "main" {
+        pretty_name.clone()
+    } else {
+        tcx.symbol_name(Instance::mono(tcx, def_id)).to_string()
+    };
+
     let loc = SourceLocation::def_id_loc(tcx, def_id);
     let file_stem = format!("{}_{mangled_name}", base_name.file_stem().unwrap().to_str().unwrap());
     let model_file = base_name.with_file_name(file_stem).with_extension(ArtifactType::SymTabGoto);