@celinval 2023-05-12
> I still need to add the test I've been working on, but the code is ready for review.

Done!