@celinval 2023-05-14
> Here's an initial set of comments. I'm having difficulty wrapping my head around how project, target crate, harness, etc are structured or related.
> 
> > The driver will now read the metadata for each crate, and collect all artifacts based on the symtab goto file that is recorded in the metadata of each harness.
> 
> Does each crate metadata file contain metadata for each harness in that crate?

Yes.That actually didn't change in this PR at all. It was a previous change.

Each crate has one metadata. Each crate metadata has a vector with harnesses metadata. Each harness metadata contains the name of a goto file.

Before this change, the name of the goto file was redundant since all harnesses metadata pointed to the same goto file. With this change, we extract the goto file for each harness from the metadata.