@celinval 2023-05-14
Btw, the crate metadata is defined here:

https://github.com/model-checking/kani/blob/8d79ee5722e47c086e1904b9f798a106bc900bf0/kani_metadata/src/lib.rs#L20-L30

It holds metadata for proof and test harnesses. In each harness metadata, we store the path to the model here:

https://github.com/model-checking/kani/blob/8d79ee5722e47c086e1904b9f798a106bc900bf0/kani_metadata/src/harness.rs#L23-L24