@celinval 2023-05-15
I manually executed s2n-quic to get the overall execution time and compilation times using a release build. I compared the head of this branch (fd0075c) and main (7c4400d):

|    |  Main | PR  
|---|--------|-------------------
|compilation time | 21s | 35s
|overall time | 12min 30s | 8min 24s

I think overall this change is still fairly beneficial for our users.  I suggest that we prioritize compiler performance improvements to a follow up PR. I created an issue for adding more data to benchcomp and to investigate optimizations.